feat: add layout for intro section
Added a header and navbar component, together these components make up
the intro section of the portfolio website. So far, a basic layout is
established with attention given to responsive design.

The `prettier-plugin-tailwindcss` package was added to allow prettier to
format strings of tailwind classes in jsx.

An `assets` folder was also added and it stores the fonts used in the
project.